+++ sharding-core/sharding-core-parse/sharding-core-parse-common/src/main/java/org/apache/shardingsphere/core/parse/parser/clause/expression/BasicExpressionParser.java
@@ -18,6 +18,7 @@
 package org.apache.shardingsphere.core.parse.parser.clause.expression;
 
 import lombok.RequiredArgsConstructor;
+import org.apache.shardingsphere.core.parse.antlr.constant.QuoteCharacter;
 import org.apache.shardingsphere.core.parse.lexer.LexerEngine;
 import org.apache.shardingsphere.core.parse.lexer.token.Literals;
 import org.apache.shardingsphere.core.parse.lexer.token.Symbol;
@@ -132,7 +133,7 @@ public final class BasicExpressionParser {
     private void setTableToken(final SQLStatement sqlStatement, final int beginPosition, final SQLPropertyExpression propertyExpr) {
         String owner = propertyExpr.getOwner().getName();
         if (sqlStatement.getTables().getTableNames().contains(SQLUtil.getExactlyValue(propertyExpr.getOwner().getName()))) {
-            sqlStatement.addSQLToken(new TableToken(beginPosition - owner.length(), 0, SQLUtil.getExactlyValue(owner), SQLUtil.getLeftDelimiter(owner), SQLUtil.getRightDelimiter(owner)));
+            sqlStatement.addSQLToken(new TableToken(beginPosition - owner.length(), 0, SQLUtil.getExactlyValue(owner), QuoteCharacter.valueFrom(SQLUtil.getStartDelimiter(owner))));
         }
     }
 }